move discord reinvite to close after matchup is over
